scripts/requirements.txt
pyeight==0.2.0
psycopg2-binary==2.9.9
